--- a/CHANGELOG
+++ b/CHANGELOG
@@ -19,6 +19,7 @@ PR - Pull Request (on GitHub), i.e. integration of a new feature or bugfix
 General:
  - speed improvements:
     - loading .gz files is about 7% faster (#8069)
+    - loading of mzML files with more than than mz/intensity (e.g. ion mobility) is 20-40% faster (#8074)
   
 Dependencies: