--- a/src/tests/class_tests/openms/source/MetaInfoInterface_test.cpp
+++ b/src/tests/class_tests/openms/source/MetaInfoInterface_test.cpp
@@ -327,6 +327,23 @@ START_SECTION((void removeMetaValue(const String& name)))
 	i.removeMetaValue("icon");
 END_SECTION
 
+START_SECTION((void addMetaValues(const MetaInfoInterface& from)))
+  MetaInfoInterface m_new, m_base;
+
+  m_base.setMetaValue("label", String("old"));
+  m_base.setMetaValue("exists_no_overwrite", 5.2);
+  
+  m_new.setMetaValue("label", String("new")); // will be overwritten
+  m_new.setMetaValue("icon", 4.3);            // will be added
+  
+  m_base.addMetaValues(m_new);
+
+  TEST_EQUAL(m_base.getMetaValue("label"), String("new"));
+  TEST_EQUAL(m_base.getMetaValue("icon"), 4.3);
+  TEST_EQUAL(m_base.getMetaValue("exists_no_overwrite"), 5.2);
+
+END_SECTION
+
 START_SECTION((void swap(MetaInfoInterface&& rhs)))
 {
   MetaInfoInterface mi1, mi2;