## Description

Faster reading of mzML when additional binary data is present (e.g. for ion mobility)
     - 20% faster on MSVC (318.4 MiB/s --> 399.8 MiB/s)
     - 40% faster on GCC (230.7 MiB/s --> 372.2 MiB/s)

Details:
avoids checking a meta-value for every(!) data point and introduces a new case when filtering by int/mz is not required (which is usually the case), allowing to copy whole arrays instead of individual items.


## Checklist
- [ ] Make sure that you are listed in the AUTHORS file
- [ ] Add relevant changes and new features to the CHANGELOG file
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] New and existing unit tests pass locally with my changes
- [ ] Updated or added python bindings for changed or new classes (Tick if no updates were necessary.)

### How can I get additional information on failed tests during CI
<details>
  <summary>Click to expand</summary>
If your PR is failing you can check out

- The details of the action statuses at the end of the PR or the "Checks" tab.
- http://cdash.seqan.de/index.php?project=OpenMS and look for your PR. Use the "Show filters" capability on the top right to search for your PR number.
  If you click in the column that lists the failed tests you will get detailed error messages.

</details>

### Advanced commands (admins / reviewer only)
<details>
  <summary>Click to expand</summary>
  
- `/reformat` (experimental) applies the clang-format style changes as additional commit. Note: your branch must have a different name (e.g., yourrepo:feature/XYZ) than the receiving branch (e.g., OpenMS:develop). Otherwise, reformat fails to push.
- setting the label "NoJenkins" will skip tests for this PR on jenkins (saves resources e.g., on edits that do not affect tests)
- commenting with `rebuild jenkins` will retrigger Jenkins-based CI builds
  
</details>

---
:warning: Note: Once you opened a PR try to minimize the number of *pushes* to it as every push will trigger CI (automated builds and test) and is rather heavy on our infrastructure (e.g., if several pushes per day are performed).


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
  - Added a new option in FileFilter to remove all binary metadata arrays (except m/z and intensity) from spectra in mzML files.

- **Documentation**
  - Updated FileFilter documentation to include new filtering options for mzML files, such as filtering by activation method, collision energy, isolation window width, zoom scans, chromatograms, metadata arrays, and empty spectra.

- **Performance Improvements**
  - Optimized metadata handling for spectra, improving efficiency when processing and inserting metadata arrays.
  - Achieved 20-40% faster loading of mzML files containing additional metadata beyond m/z and intensity.

- **Refactor**
  - Streamlined and clarified internal metadata merging and handling logic for better maintainability.
  - Consolidated metadata processing into a single, more efficient code path, removing redundant helper functions.

- **Tests**
  - Added tests verifying meta value merging behavior, ensuring proper overwriting and addition of metadata entries.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->